[3D Cuboid testing] Built

+++ orb_slam3_ros/orb_slam3/src/Optimizer.cc
@@ -2258,7 +2258,7 @@ void Optimizer::LocalBundleAdjustment(KeyFrame *pKF, bool* pbStopFlag, Map* pMap
             if (pMP->Observations() == 1)
                 continue;
             g2o::VertexSBAPointXYZ *vPoint = static_cast<g2o::VertexSBAPointXYZ *>(optimizer.vertex(pMP->mnId + maxKFid + maxObjectid + 2));
-            pMP->SetWorldPos(vPoint->estimate());
+            pMP->SetWorldPos(vPoint->estimate().cast<float>());
             pMP->UpdateNormalAndDepth();
         }
